Healthy-IFR/python-app/requirements.txt
pillow
torch
torchvision
# flask==2.1.0